+++ packages/sdk-react/src/index.ts
@@ -39,6 +39,8 @@ export type {
   TicketTransfer,
   TicketTransferStatus,
   TicketTransfersListResponse,
+  TicketTransferTicketActionResult,
+  TicketTransferTicketArgs,
   TokenResponse,
   Transaction,
   TransactionLineItem,